+++ api/graphql/resolvers/project.utils.js
@@ -1,5 +1,5 @@
 const { filter, find } = require('underscore')
-const value = require('get-value')
+const get = require('get-value')
 
 /**
  * Connects projects, customers and labels
@@ -14,7 +14,7 @@ function connectEntities(projects, customers, labels) {
       ...project,
       customer: find(customers, c => c.key === project.customerKey),
       labels: filter(labels, lbl => {
-        const str = value(project, 'labels', { default: '' })
+        const str = get(project, 'labels', { default: '' })
         return str.indexOf(lbl.name) !== -1
       }),
     }))