+++ api/graphql/resolvers/timesheet.js
@@ -1,10 +1,9 @@
-const { first, filter, find, pick, contains } = require('underscore')
+const { filter, find, pick, contains, isEmpty } = require('underscore')
 const { formatDate, getMonthIndex, getWeek } = require('../../../utils')
 const EventMatching = require('./timesheet.matching')
 const { connectEntities } = require('./project.utils')
 const { getPeriods, connectTimeEntries } = require('./timesheet.utils')
 const value = require('get-value')
-const log = require('debug')('api/graphql/resolvers/timesheet')
 const { gql, AuthenticationError, ApolloError } = require('apollo-server-express')
 
 const typeDef = gql`
@@ -45,8 +44,7 @@ const typeDef = gql`
     isConfirmed: Boolean
     isForecasted: Boolean
     isForecast: Boolean
-    confirmedDuration: Float!
-    forecastedDuration: Float!
+    forecastedHours: Float!
   }
 
   """
@@ -66,7 +64,7 @@ const typeDef = gql`
     startDateTime: String!
     endDateTime: String!
     matchedEvents: [EventInput]
-    isForecast: Boolean
+    forecastedHours: Float
   }
 
   extend type Query {
@@ -80,15 +78,22 @@ const typeDef = gql`
     """
     Adds matched time entries for the specified period and an entry for the confirmed period
     """
-    submitPeriod(period: TimesheetPeriodInput!): BaseResult!
+    submitPeriod(period: TimesheetPeriodInput!, forecast: Boolean!): BaseResult!
 
     """
     Deletes time entries for the specified period and the entry for the confirmed period
     """
-    unsubmitPeriod(period: TimesheetPeriodInput!): BaseResult!
+    unsubmitPeriod(period: TimesheetPeriodInput!, forecast: Boolean!): BaseResult!
   }
 `
 
+/**
+ * Timesheet
+ * 
+ * @param {*} _obj {}
+ * @param {*} variables Variables: startDateTime, endDateTime, dateFormat, locale
+ * @param {*} ctx GraphQL context
+ */
 async function timesheet(_obj, variables, ctx) {
   if (!ctx.services.msgraph) throw new AuthenticationError()
   try {
@@ -114,7 +119,12 @@ async function timesheet(_obj, variables, ctx) {
 
     for (let i = 0; i < periods.length; i++) {
       let period = periods[i]
-      let confirmed = await ctx.services.azstorage.getConfirmedPeriod(ctx.user.id, period.id)
+      let [confirmed, forecasted] = await Promise.all([
+        ctx.services.azstorage.getConfirmedPeriod(ctx.user.id, period.id),
+        ctx.services.azstorage.getForecastedPeriod(ctx.user.id, period.id)
+      ])
+      period.isForecasted = !!forecasted
+      period.forecastedHours = period.isForecasted && forecasted.hours
       period.isConfirmed = !!confirmed
       if (period.isConfirmed) {
         period.events = connectTimeEntries(
@@ -124,23 +134,17 @@ async function timesheet(_obj, variables, ctx) {
           labels,
         )
         period.matchedEvents = period.events
-        period.confirmedDuration = confirmed.hours
       } else {
-        if (period.isForecast) {
-          let forecasted = await ctx.services.azstorage.getForecastedPeriod(ctx.user.id, period.id)
-          period.isForecasted = !!forecasted
-          if (period.isForecasted) {
-            let timeentries = await ctx.services.azstorage.getTimeEntries(
-              {
-                resourceId: ctx.user.id,
-                startDateTime: variables.startDateTime,
-                endDateTime: variables.endDateTime,
-              },
-              { sortAsc: true, forecast: true }
-            )
-            period.events = connectTimeEntries(timeentries, projects, customers, labels)
-            period.forecastedDuration = forecasted.hours
-          }
+        if (period.isForecast && period.isForecasted) {
+          let timeentries = await ctx.services.azstorage.getTimeEntries(
+            {
+              resourceId: ctx.user.id,
+              startDateTime: variables.startDateTime,
+              endDateTime: variables.endDateTime,
+            },
+            { sortAsc: true, forecast: true }
+          )
+          period.events = connectTimeEntries(timeentries, projects, customers, labels)
         }
         if (!period.events) {
           period.events = await ctx.services.msgraph.getEvents(period.startDateTime, period.endDateTime)
@@ -159,34 +163,50 @@ async function timesheet(_obj, variables, ctx) {
   }
 }
 
+/**
+ * Submit period
+ * 
+ * @param {*} _obj {}
+ * @param {*} variables Variables: period, forecast
+ * @param {*} ctx GraphQL context
+ */
 async function submitPeriod(_obj, variables, ctx) {
   try {
-    let hours = 0
-    if (variables.period.matchedEvents.length > 0) {
+    let { period, forecast } = { ...variables }
+    period.hours = 0
+    console.log(period)
+    if (!isEmpty(period.matchedEvents)) {
       const [events, labels] = await Promise.all([
-        ctx.services.msgraph.getEvents(variables.period.startDateTime, variables.period.endDateTime),
+        ctx.services.msgraph.getEvents(period.startDateTime, period.endDateTime),
         ctx.services.azstorage.getLabels(),
       ])
-
-      let timeentries = variables.period.matchedEvents
-        .map(entry => {
-          const event = find(events, e => e.id === entry.id)
-          if (!event) return
-          const _labels = filter(labels, lbl => contains(event.categories, lbl.name)).map(lbl => lbl.name)
-          return {
-            user: ctx.user,
-            entry,
-            event,
-            labels: _labels,
-          }
-        })
-        .filter(entry => entry)
-      hours = await ctx.services.azstorage.addTimeEntries(variables.period.id, timeentries, variables.period.isForecast)
+      let timeentries = period.matchedEvents.reduce((arr, event) => {
+        const entry = {
+          ...pick(event, 'projectId', 'manualMatch'),
+          event: find(events, e => e.id === event.id),
+        }
+        if (!entry.event) return arr
+        entry.labels = filter(labels, lbl => contains(event.categories, lbl.name)).map(lbl => lbl.name)
+        return [...arr, entry]
+      }, [])
+      period.hours = await ctx.services.azstorage.addTimeEntries(
+        pick(period, 'id'),
+        ctx.user,
+        timeentries,
+        forecast
+      )
     }
-    if (variables.period.isForecast) {
-      await ctx.services.azstorage.addForecastedPeriod(variables.period.id, ctx.user.id, hours)
-    } else {
-      await ctx.services.azstorage.addConfirmedPeriod(variables.period.id, ctx.user.id, hours)
+    if (forecast) {
+      await ctx.services.azstorage.addForecastedPeriod(
+        pick(period, 'id', 'hours'),
+        ctx.user.id
+      )
+    }
+    else {
+      await ctx.services.azstorage.addConfirmedPeriod(
+        pick(period, 'id', 'hours', 'forecastedHours'),
+        ctx.user.id
+      )
     }
     return { success: true, error: null }
   } catch (error) {
@@ -197,14 +217,25 @@ async function submitPeriod(_obj, variables, ctx) {
   }
 }
 
+/**
+ * Unsubmit period
+ * 
+ * @param {*} _obj {}
+ * @param {*} variables Variables: period, forecast
+ * @param {*} ctx GraphQL context
+ */
 async function unsubmitPeriod(_obj, variables, ctx) {
   try {
-    if (variables.period.isForecast) {
-      await ctx.services.azstorage.deleteTimeEntries(variables.period.id, ctx.user.id, true)
-      await ctx.services.azstorage.removeForecastedPeriod(variables.period.id, ctx.user.id)
+    if (variables.forecast) {
+      await Promise.all([
+        ctx.services.azstorage.deleteTimeEntries(variables.period.id, ctx.user.id, true),
+        ctx.services.azstorage.removeForecastedPeriod(variables.period.id, ctx.user.id)
+      ])
     } else {
-      await ctx.services.azstorage.deleteTimeEntries(variables.period.id, ctx.user.id, false)
-      await ctx.services.azstorage.removeConfirmedPeriod(variables.period.id, ctx.user.id)
+      await Promise.all([
+        ctx.services.azstorage.deleteTimeEntries(variables.period.id, ctx.user.id, false),
+        ctx.services.azstorage.removeConfirmedPeriod(variables.period.id, ctx.user.id)
+      ])
     }
     return { success: true, error: null }
   } catch (error) {