+++ client/pages/Timesheet/ActionBar/commands.tsx
@@ -1,6 +1,7 @@
 import { DefaultButton, IButtonStyles, PrimaryButton } from 'office-ui-fabric-react/lib/Button'
-import { IContextualMenuItem } from 'office-ui-fabric-react/lib/ContextualMenu'
+import { IContextualMenuItem, IContextualMenuProps } from 'office-ui-fabric-react/lib/ContextualMenu'
 import * as React from 'react'
+import { first, omit } from 'underscore'
 import { ITimesheetContext } from '../context'
 import styles from './ActionBar.module.scss'
 import { ACTIONBAR_ICON_PROPS } from './ACTIONBAR_ICON_PROPS'
@@ -8,8 +9,8 @@ import { WeekPicker } from './WeekPicker'
 
 const buttonStyles: IButtonStyles = { root: { height: 44, marginLeft: 4 } }
 
-export const WEEK_PICKER = ({ selectedPeriod, periods, t }: ITimesheetContext): IContextualMenuItem => ({
-    key: 'WEEK_PICKER',
+export const WEEK_PICKER_COMMAND = ({ selectedPeriod, periods, t }: ITimesheetContext): IContextualMenuItem => ({
+    key: 'WEEK_PICKER_COMMAND',
     onRender: () => (
         <>
             <WeekPicker />
@@ -22,8 +23,8 @@ export const WEEK_PICKER = ({ selectedPeriod, periods, t }: ITimesheetContext): 
     ),
 })
 
-export const GO_TO_CURRENT_WEEK = ({ scope, dispatch, t }: ITimesheetContext): IContextualMenuItem => ({
-    key: 'GO_TO_CURRENT_WEEK',
+export const GO_TO_CURRENT_WEEK_COMMAND = ({ scope, dispatch, t }: ITimesheetContext): IContextualMenuItem => ({
+    key: 'GO_TO_CURRENT_WEEK_COMMAND',
     iconOnly: true,
     iconProps: { iconName: 'RenewalCurrent', ...ACTIONBAR_ICON_PROPS },
     onClick: () => dispatch({ type: 'MOVE_SCOPE', payload: new Date().toISOString() }),
@@ -31,26 +32,26 @@ export const GO_TO_CURRENT_WEEK = ({ scope, dispatch, t }: ITimesheetContext): I
     title: t('timesheet.goToCurrentWeek'),
 })
 
-export const GO_TO_PREV_WEEK = ({ dispatch, t }: ITimesheetContext): IContextualMenuItem => ({
-    key: 'GO_TO_PREV_WEEK',
+export const GO_TO_PREV_WEEK_COMMAND = ({ dispatch, t }: ITimesheetContext): IContextualMenuItem => ({
+    key: 'GO_TO_PREV_WEEK_COMMAND',
     iconOnly: true,
     iconProps: { iconName: 'Back', ...ACTIONBAR_ICON_PROPS },
     onClick: () => dispatch({ type: 'MOVE_SCOPE', payload: { amount: -1, unit: 'week' } }),
     title: t('timesheet.goToPrevWeek')
 })
 
-export const GO_TO_NEXT_WEEK = ({ dispatch, t }: ITimesheetContext): IContextualMenuItem => ({
-    key: 'GO_TO_NEXT_WEEK',
+export const GO_TO_NEXT_WEEK_COMMAND = ({ dispatch, t }: ITimesheetContext): IContextualMenuItem => ({
+    key: 'GO_TO_NEXT_WEEK_COMMAND',
     iconOnly: true,
     iconProps: { iconName: 'Forward', ...ACTIONBAR_ICON_PROPS },
     onClick: () => dispatch({ type: 'MOVE_SCOPE', payload: { amount: 1, unit: 'week' } }),
     title: t('timesheet.goToNextWeek'),
 })
 
-export const SELECT_PERIOD = ({ periods, loading, selectedPeriod, dispatch, t }: ITimesheetContext): IContextualMenuItem[] => {
+export const SELECT_PERIOD_COMMANDS = ({ periods, loading, selectedPeriod, dispatch, t }: ITimesheetContext): IContextualMenuItem[] => {
     if (periods.length === 1) return []
     return periods.map((period, idx) => ({
-        key: `SELECT_PERIOD_${idx}`,
+        key: `SELECT_PERIOD_COMMANDS_${idx}`,
         onRender: () => (
             <DefaultButton
                 hidden={!!loading}
@@ -64,51 +65,103 @@ export const SELECT_PERIOD = ({ periods, loading, selectedPeriod, dispatch, t }:
     }))
 }
 
-export const CONFIRM_ACTIONS = (context: ITimesheetContext): IContextualMenuItem => ({
-    key: 'CONFIRM_ACTIONS',
+export const CONFIRM_FORECAST_COMMANDS = (context: ITimesheetContext): IContextualMenuItem => ({
+    key: 'CONFIRM_FORECAST_COMMANDS',
     onRender: () => {
-        if (context.selectedPeriod.isForecast || context.loading || !!context.error) return null
-        if (context.selectedPeriod.isConfirmed) {
-            return (
-                <DefaultButton
-                    iconProps={{ iconName: 'Cancel' }}
-                    onClick={context.onUnsubmitPeriod}
-                    text={context.t('timesheet.unconfirmHoursText')}
-                    className={styles.unconfirmButton}
-                    styles={buttonStyles} />
-            )
+        if (context.loading || !!context.error) return null
+        const {
+            isComplete,
+            isForecast,
+            isForecasted,
+            isConfirmed,
+            isPast,
+        } = context.selectedPeriod
+        const commandProps = {
+            FORECAST_PERIOD: {
+                key: 'FORECAST_PERIOD',
+                styles: buttonStyles,
+                iconProps: { iconName: 'BufferTimeBefore' },
+                onClick: () => context.onSubmitPeriod(true),
+                canCheck: true,
+                text: context.t('timesheet.forecastHoursText'),
+                secondaryText: context.t('timesheet.forecastHoursSecondaryText'),
+            },
+            UNFORECAST_PERIOD: {
+                key: 'UNFORECAST_PERIOD',
+                styles: buttonStyles,
+                iconProps: { iconName: 'Cancel' },
+                onClick: () => context.onUnsubmitPeriod(true),
+                canCheck: true,
+                text: context.t('timesheet.unforecastHoursText'),
+                secondaryText: context.t('timesheet.unforecastHoursSecondaryText'),
+            },
+            CONFIRM_PERIOD: {
+                key: 'CONFIRM_PERIOD',
+                className: styles.confirmPeriodButton,
+                styles: buttonStyles,
+                iconProps: { iconName: 'CheckMark' },
+                onClick: () => context.onSubmitPeriod(false),
+                canCheck: true,
+                text: context.t('timesheet.confirmHoursText'),
+                secondaryText: context.t('timesheet.confirmHoursSecondaryText')
+            },
+            UNCONFIRM_PERIOD: {
+                key: 'UNCONFIRM_PERIOD',
+                className: styles.unconfirmPeriodButton,
+                styles: buttonStyles,
+                iconProps: { iconName: 'Cancel' },
+                onClick: () => context.onUnsubmitPeriod(false),
+                canCheck: true,
+                text: context.t('timesheet.unconfirmHoursText'),
+                secondaryText: context.t('timesheet.unconfirmHoursSecondaryText')
+            }
         }
-        return (
-            <PrimaryButton
-                disabled={!context.selectedPeriod.isComplete}
-                iconProps={{ iconName: 'CheckMark' }}
-                onClick={context.onSubmitPeriod}
-                text={context.t('timesheet.confirmHoursText')}
-                className={styles.confirmButton}
-                styles={buttonStyles} />
-        )
-    }
-})
 
-export const FORECAST_ACTIONS = (context: ITimesheetContext): IContextualMenuItem => ({
-    key: 'FORECAST_ACTIONS',
-    onRender: () => {
-        if (!context.selectedPeriod.isForecast || context.loading || !!context.error) return null
-        if (context.selectedPeriod.isForecasted) {
-            return (
-                <DefaultButton
-                    iconProps={{ iconName: 'Cancel' }}
-                    onClick={context.onUnsubmitPeriod}
-                    text={context.t('timesheet.unforecastHoursText')}
-                    styles={buttonStyles} />
-            )
+        const commands = []
+
+        if (isConfirmed) commands.push(commandProps.UNCONFIRM_PERIOD)
+        else if (isForecast) {
+            if (isForecasted) {
+                if (isComplete) commands.push(commandProps.CONFIRM_PERIOD)
+                commands.push(commandProps.UNFORECAST_PERIOD)
+            }
+            else {
+                commands.push(commandProps.FORECAST_PERIOD)
+                commands.push(commandProps.CONFIRM_PERIOD)
+            }
         }
+        else {
+            if (isComplete) {
+                commands.push(commandProps.CONFIRM_PERIOD)
+                if (!isPast) {
+                    if (isForecasted) commands.push(commandProps.UNFORECAST_PERIOD)
+                    else commands.push(commandProps.FORECAST_PERIOD)
+                }
+            }
+            else {
+                if (!isPast) {
+                    if (isForecasted) commands.push(commandProps.UNFORECAST_PERIOD)
+                    else commands.push(commandProps.FORECAST_PERIOD)
+                }
+                commands.push({ ...commandProps.CONFIRM_PERIOD, disabled: true })
+            }
+        }
+        
+        let menuProps: IContextualMenuProps = null
+        if (commands.length > 1) {
+            menuProps = {
+                items: commands.map((cmd, idx) => ({
+                    ...omit(cmd, 'buttonStyles', 'iconProps'),
+                    isChecked: idx === 0
+                }))
+            }
+        }
+
         return (
             <PrimaryButton
-                iconProps={{ iconName: 'BufferTimeBefore' }}
-                onClick={context.onSubmitPeriod}
-                text={context.t('timesheet.forecastHoursText')}
-                styles={buttonStyles} />
+                primary={false}
+                {...first(commands)}
+                menuProps={menuProps} />
         )
     }
 })