+++ client/pages/Timesheet/Overview/StatusBar/index.tsx
@@ -66,7 +66,7 @@ export const StatusBar = () => {
                 'timesheet.periodForecastedText',
                 { hours: dateUtils.getDurationString(context.selectedPeriod.matchedDuration, t) }
             ),
-            type: MessageBarType.success,
+            type: MessageBarType.info,
             iconName: 'BufferTimeBoth'
         })
     }